+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- 📝 Fixed string version of standard output from commands #47
 - ⬆️🐘 Gradle (4.7) #46
 - ⬆️ Kotlin (1.2.41) #46
 - ⬆️ JUnit Platform (1.2.0) #46